Updated Masterchef.sol
We removed the part where we tried to implement the remove_pool feature of the farm pools. Now we are sticking with the idea that, even if we set the allocation points for a pool to be zero, the users' tokens would be safe. And they would be able to withdraw them anytime they wish.